feat: 去除七牛云/引入腾讯云

+++ src/init/initFFmpeg.ts
@@ -1,13 +1,10 @@
 import { execSync, spawnSync } from 'child_process';
 
-import { getRangeRandom } from 'billd-utils';
-
 import { PROJECT_ENV, PROJECT_ENV_ENUM } from '@/constant';
+import { initUser } from '@/init/initData';
 import liveService from '@/service/live.service';
 import { chalkERROR, chalkSUCCESS, chalkWARN } from '@/utils/chalkTip';
-import { qiniuUtils } from '@/utils/qiniu';
-
-import { initUser } from './initData';
+import { tencentcloudUtils } from '@/utils/tencentcloud';
 
 function ffmpegIsInstalled() {
   const res = spawnSync('ffmpeg', ['-version']);
@@ -30,40 +27,42 @@ async function addLive({
   flvurl: string;
   base64: string;
 }) {
-  const isExist = await qiniuUtils.queryAFlow({ roomId: live_room_id });
-  if (!isExist) {
-    await qiniuUtils.createAFlow({ roomId: live_room_id });
-  }
-  const remoteFlv = qiniuUtils.generateRtmpPublishUrl({ roomId: live_room_id });
-  const flvurl = qiniuUtils.getFlvPullUrl({ roomId: live_room_id });
-  // ffmpeg后台运行
-  // https://www.jianshu.com/p/6ea70e6d8547
-  // 1 代表标准输出
-  // 2 代表标准错误
-  // 1>/dev/null 把标准输出导入到null设备,也就是消失不见，如果要重定向到某个文件，可以1>1.txt
-  // 2>&1 把标准错误也导入到标准输出同样的地方
-  // -loglevel quiet不输出log
-  // const ffmpeg = `ffmpeg -loglevel quiet -stream_loop -1 -re -i ${localFile} -c copy -f flv '${remoteFlv}' 1>/dev/null 2>&1 &`;
-  const ffmpeg = `ffmpeg -stream_loop -1 -re -i ${localFile} -c copy -f flv '${remoteFlv}' 1>/dev/null 2>&1 &`;
-  // const ffmpeg = `echo test initFFmpeg`;
-  setTimeout(() => {
-    console.log('推流', ffmpeg);
-    execSync(ffmpeg);
-  }, getRangeRandom(1000, 5000));
-  console.log('ffmpeg命令', ffmpeg);
-  const socketId = live_room_id;
-  await liveService.deleteBySocketId(socketId);
-  await liveService.create({
-    live_room_id,
-    user_id,
-    socketId,
-    system: 1,
-    track_audio: true,
-    track_video: true,
-    coverImg: base64,
-    streamurl: '',
-    flvurl,
+  const { res, err } = await tencentcloudUtils.queryLiveStream({
+    roomId: live_room_id,
   });
+  if (err) {
+    return;
+  }
+  if (res) {
+    const remoteFlv = tencentcloudUtils.getPushUrl({
+      roomId: live_room_id,
+    });
+    // ffmpeg后台运行
+    // https://www.jianshu.com/p/6ea70e6d8547
+    // 1 代表标准输出
+    // 2 代表标准错误
+    // 1>/dev/null 把标准输出导入到null设备,也就是消失不见，如果要重定向到某个文件，可以1>1.txt
+    // 2>&1 把标准错误也导入到标准输出同样的地方
+    // -loglevel quiet不输出log
+    const ffmpeg = `ffmpeg -loglevel quiet -stream_loop -1 -re -i ${localFile} -c copy -f flv '${remoteFlv}' 1>/dev/null 2>&1 &`;
+    // const ffmpeg = `echo test initFFmpeg`;
+    console.log('ffmpeg命令', ffmpeg);
+    execSync(ffmpeg);
+    const flvurl = tencentcloudUtils.getPullUrl({ roomId: live_room_id }).flv;
+    const socketId = live_room_id;
+    await liveService.deleteBySocketId(socketId);
+    await liveService.create({
+      live_room_id,
+      user_id,
+      socketId,
+      system: 1,
+      track_audio: true,
+      track_video: true,
+      coverImg: base64,
+      streamurl: '',
+      flvurl,
+    });
+  }
 }
 
 export const initFFmpeg = async (init = true) => {
